refactor(auth): Reogranizing the folder structure of test and doc to the API path standard
